feat: modify vtparse to be generic
There was an original part of the code (user_data) that we didn't port
over to our vtparse implementation. This makes our implementation
incomplete and hard to work with. However user_data was a void*. With
that in mind, making a generic implementation that allows us to now
play with our own data while using the callback will hopefully work
quite well.